pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cekdata-engine"
version = "0.1.0"
description = "Mesin verifikasi klaim berbasis data statistik Indonesia"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["fact-checking", "data-verification", "indonesia", "statistics", "journalism"]

dependencies = [
    # AI analysis — opsional, fallback tersedia jika tidak terpasang
    "openai>=1.30",
    # Vector search backend — opsional
    "pinecone-client>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
cekdata = "cekdata_engine.cli:main"

[tool.setuptools.packages.find]
include = ["cekdata_engine*"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# ---------------------------------------------------------------------------
# ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long — dihandle oleh formatter
    "B006",  # mutable default args — ada kasus sah di dataclass field
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # assert diizinkan di test

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true   # indicator_registry & breakdown_registry belum punya stubs
